pyproject.toml
[tool.poetry]
name = "mslab-discord-bot"
version = "0.1.0"
description = ""
authors = ["kjchan <r11922032@csie.ntu.edu.tw>"]
readme = "README.md"
packages = [{include = "mslab_discord_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
black = "^23.7.0"
python-dotenv = "^1.0.0"
dokuwiki = "^1.3.3"
discord = "^2.3.2"
aiocron = "^1.8"
pytest = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
